requirements.txt
Flask
PyMuPDF
pdf2image
pytesseract
python-docx
Pillow
